feat(core): 优化消息重试机制和线程池配置
- 重发/重试消息接收并发数统一为 retryConcurrency
- 新增 pollBlockMillis 配置项，用于设置 Redis Stream轮询阻塞时间
- 新增 deliverGroupThreadPoolSize 和 deliverGroupThreadKeepAliveTime 配置项，用于设置消息投递线程池配置
- 优化消息轮询和投递逻辑，提高系统稳定性和性能

+++ eventbus-core/src/main/java/com/github/likavn/eventbus/core/utils/GroupedThreadPoolExecutor.java
@@ -57,7 +57,7 @@ public class GroupedThreadPoolExecutor {
     private final AtomicInteger currentCorePoolSize = new AtomicInteger(0);
 
     private final int corePoolSize;
-    private final int keepAliveTimeMillis;
+    private final long keepAliveTimeMillis;
     private final NamedThreadFactory threadFactory;
 
     /**
@@ -72,7 +72,7 @@ public class GroupedThreadPoolExecutor {
     @Setter
     private Consumer<WorkerThread> afterConsumer;
 
-    public GroupedThreadPoolExecutor(int corePoolSize, int keepAliveTimeMillis, NamedThreadFactory threadFactory) {
+    public GroupedThreadPoolExecutor(int corePoolSize, long keepAliveTimeMillis, NamedThreadFactory threadFactory) {
         this.corePoolSize = corePoolSize;
         this.keepAliveTimeMillis = keepAliveTimeMillis;
         this.threadFactory = threadFactory;
